+++ nodeservices/server.js
@@ -1,33 +1,15 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-let User_Registration = require('./modelclasses/user_registration.js'); // model class import
 const mongoose = require('mongoose');
-const uuid = require('uuid/v4'); // generate UUID
-const PropertiesReader = require('properties-reader'); // read from properties file
-const prop = PropertiesReader('url.properties');
-
+let utils = require('./utils/Utils.js');
+const conString = utils.getProperty('mongo_connect_url');
+let user_registration_api = require('./apis/UserRegistrationAPI.js');
 
 let app = express();
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(express.static(__dirname))
 
-// -------------- utils ------------------>
-
-getProperty = (pty) => {
-  return prop.get(pty);
-}
-const conString = getProperty('mongo_connect_url');
-
-generateResponse = (status,errorCode,errorDesc) => {
-  var respObj = {};
-  respObj.status = status;
-  respObj.errorCode = errorCode;
-  respObj.errorDesc = errorDesc;
-
-  return respObj;
-
-}
 
 mongoose.connect(conString, { useNewUrlParser: true }, (err) => {
   console.log("Database connection", err)
@@ -42,7 +24,7 @@ app.use(function(req, res, next) {
 });
 
 // -------------- server port listener ------------------>
-app.listen(getProperty('server.port'), () => {
+app.listen(utils.getProperty('server.port'), () => {
   console.log("Well done, now I am listening...")
 })
 
@@ -50,44 +32,16 @@ app.listen(getProperty('server.port'), () => {
 
 // -------------- post user registration ------------------>
 
-app.post(getProperty('user_registration_url'),function(req,res){
-  dupCheck(req,res);
+app.post(utils.getProperty('user_registration_url'),function(req,res){
+  user_registration_api.dupCheck(req,res);
 })
 
 
-let dupCheck = function(req,res){
-  User_Registration.find({
-    $or:[{
-        'mobile': req.body.mobile
-      },{
-        'emailId': req.body.emailId
-      }]
-  },function (err,results) {
-    console.log(results.length);
-    if(results.length == 0){
-        registerData(req,res);
-    }else
-      res.send(generateResponse(getProperty("failure"),"", getProperty("user_exists")));
-  });
-}
-
-let registerData = function(req,res){
- var input = new User_Registration(req.body);
-  input.profileId = uuid();
-  input.save(function(error){
-    if(error){
-      console.log(error.errors);
-      res.send(generateResponse(getProperty("failure"), getProperty("invalid_input_code"),getProperty("invalid_input")));
-    }else{
-      res.send(generateResponse(getProperty("success"), getProperty("reg_success_code"),getProperty("user_registered")));
-    }
-  })
-
-}
+// --------------- ping service --------------------------->
 
 
 app.get('/',function(req,res){
- res.send("Hello");
+  res.send("Hello");
 })
